requirements.txt
argon2-cffi
cryptography
sqlalchemy
prompt_toolkit
